# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: EarnQA API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /questions:
    get:
      operationId: listQuestions
      tags: [questions]
      summary: List questions
      parameters:
        - name: category
          in: query
          schema:
            type: ["string", "null"]
        - name: type
          in: query
          schema:
            type: ["string", "null"]
        - name: status
          in: query
          schema:
            type: ["string", "null"]
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
        - name: offset
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List of questions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListQuestionsResponse"
    post:
      operationId: createQuestion
      tags: [questions]
      summary: Create a custom question (costs 100 cents)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateQuestionBody"
      responses:
        "201":
          description: Question created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "402":
          description: Insufficient funds
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /questions/{id}:
    get:
      operationId: getQuestion
      tags: [questions]
      summary: Get a question by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Question details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuestionWithStats"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    Question:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        type:
          type: string
          enum: [short_answer, poll, rating]
        category:
          type: string
        status:
          type: string
          enum: [active, pending, rejected]
        pollOptions:
          type: ["array", "null"]
          items:
            type: string
        creatorId:
          type: ["string", "null"]
        creatorName:
          type: ["string", "null"]
        isCustom:
          type: boolean
        totalAnswers:
          type: integer
        rejectionReason:
          type: ["string", "null"]
        rejectedAt:
          type: ["string", "null"]
        rejectedBy:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - title
        - type
        - category
        - status
        - isCustom
        - totalAnswers
        - createdAt
    QuestionWithStats:
      allOf:
        - $ref: "#/components/schemas/Question"
        - type: object
          properties:
            ratingAverage:
              type: ["number", "null"]
            pollResults:
              type: ["array", "null"]
              items:
                type: object
                properties:
                  option:
                    type: string
                  count:
                    type: integer
                  percentage:
                    type: number
                required:
                  - option
                  - count
                  - percentage
            userHasAnswered:
              type: boolean
          required:
            - userHasAnswered
    ListQuestionsResponse:
      type: object
      properties:
        questions:
          type: array
          items:
            $ref: "#/components/schemas/Question"
        total:
          type: integer
      required:
        - questions
        - total
    CreateQuestionBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: ["string", "null"]
        type:
          type: string
          enum: [short_answer, poll, rating]
        category:
          type: string
        pollOptions:
          type: ["array", "null"]
          items:
            type: string
      required:
        - title
        - type
        - category
...